pyproject.toml
[project]
name = "gitlabbackup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]
